+++ verif/tests/elements/shell/test_shell_mitc4_08.py
@@ -88,7 +88,7 @@ lp0= lPatterns.newLoadPattern("default","0")
 f1= preprocessor.getSets.getSet("f1")
 nNodes= f1.getNumNodes
  
-node= f1.getNodeIJK(1,NumDivI/2+1,NumDivJ/2+1)
+node= f1.getNodeIJK(1, int(NumDivI/2+1), int(NumDivJ/2+1))
 # print("Central node: ", node.tag)
 # print("Central node coordinates: ", node.getCoo)
 lp0.newNodalLoad(node.tag,xc.Vector([0,0,-ptLoad,0,0,0])) # Concentrated load
@@ -107,7 +107,7 @@ f1= preprocessor.getSets.getSet("f1")
 
 nodes= preprocessor.getNodeHandler
 
-node= f1.getNodeIJK(1,NumDivI/2+1,NumDivJ/2+1)
+node= f1.getNodeIJK(1, int(NumDivI/2+1), int(NumDivJ/2+1))
 # print("Central node: ", node.tag)
 # print("Central node coordinates: ", node.getCoo)
 # print("Central node displacements: ", node.getDisp)
