+++ verif/tests/elements/shell/test_shell_mitc4_16.py
@@ -108,7 +108,7 @@ f1= preprocessor.getSets.getSet("f1")
 
 nodes= preprocessor.getNodeHandler
 
-node= f1.getNodeIJK(1,NumDivI/2+1,NumDivJ/2+1)
+node= f1.getNodeIJK(1, int(NumDivI/2+1), int(NumDivJ/2+1))
 # print("Central node: ", node.tag)
 # print("Central node coordinates: ", node.getCoo)
 # print("Central node displacements: ", node.getDisp)
